Revert "Move wallet types to Wallet class"
This reverts commit fdb5cdaa408297631526ec37d5208b8c5f5a3db1.
