+++ src/components/importWallet/ImportFromSeed.vue
@@ -52,10 +52,8 @@ import { mapActions } from 'vuex';
 import PasswordModal from '@/components/modal/PasswordModal';
 import modalMixin from '@/mixins/modal';
 import formMixin from '@/mixins/form';
-import { Wallet } from '@/class';
+import { WALLET_TYPE } from '@/constants';
 import WalletsList from './WalletsList';
-
-const WALLET_TYPE = Wallet.getTypes();
 
 export default {
   name: 'ImportFromSeed',
