+++ src/store/accounts/actions.js
@@ -12,6 +12,7 @@ import {
   web3,
 } from '@/class';
 import { keystore } from '@endpass/utils';
+import { WALLET_TYPE } from '@/constants';
 import {
   CHANGE_INIT_STATUS,
   SET_ADDRESS,
@@ -20,8 +21,6 @@ import {
   SET_BALANCE,
   SET_HD_CACHE_BY_TYPE,
 } from './mutations-types';
-
-const WALLET_TYPE = Wallet.getTypes();
 
 const { toChecksumAddress, fromWei } = web3.utils;
