+++ src/store/tokens/actions.js
@@ -1,3 +1,4 @@
+import { mapValues } from 'lodash';
 import {
   SET_LOADING,
   SET_TOKENS_BY_ADDRESS,
