+++ test/unit/specs/components/importWallet/WalletsList/WalletsList.spec.js
@@ -7,11 +7,9 @@ import UIComponents from '@endpass/ui';
 import { wrapShallowMountFactory } from '@/testUtils';
 
 import WalletsList from '@/components/importWallet/WalletsList/WalletsList';
-import { Wallet } from '@/class';
+import { WALLET_TYPE } from '@/constants';
 
 import { addresses, address } from 'fixtures/accounts';
-
-const WALLET_TYPE = Wallet.getTypes();
 
 const addressPos = addresses.indexOf(address);
